sellCurve tests fixed to take into account q - 1. Tests pass now.

+++ src/liquidity/calculators/ProtocolNFTAMMCalcLinear.sol
@@ -15,7 +15,7 @@ contract ProtocolNFTAMMCalcLinear is IProtocolAMMFactoryCalculator, CurveErrors 
 
     using Curve for Line;
 
-    uint8 constant PRECISION_DECIMALS = 8;
+    uint256 constant PRECISION_DECIMALS = 8;
     uint256 constant ATTO = 10 ** 18;
     uint256 constant Y_MAX = 1_000_000_000_000_000_000_000_000 * ATTO;
     uint256 constant M_MAX = 1_000_000_000_000_000_000_000_000 * 10 ** PRECISION_DECIMALS;
